feat(fork): accept setup macro parameter
This change allows specifying as a macro parameter
a setup function to change environment and possibly
other properties of the Command to be executed.

It is sometimes desirable to set additional environment
variables to child processes before running them,
e.g. when testing dynamically-loaded cdylibs and
needing to set the library search path to a folder
within the project.

Additionally, it solves a few compiler warnings
and bumps the Rust edition to 2021.